Reformatted Repository Structure
