+++ .gitignore
@@ -1,0 +1,4 @@
+# KiCad
+*.bak
+*.sch-bak
+schematics/KiCad/8-bit-cpu-schematics-cache.lib
