--- Pulse Measure
@@ -1,38 +1,0 @@
-#define pulse_ip 7
-int ontime,offtime,duty;
-float freq,period;
-   
-void setup()
-{
-  Serial.begin(9600);
-  pinMode(pulse_ip,INPUT);
-}
-void loop()
-{
-  ontime = pulseIn(pulse_ip,HIGH);
-  offtime = pulseIn(pulse_ip,LOW);
-  period = ontime+offtime;
-  freq = 1000000.0/period;
-  duty = (ontime/period)*100; 
-  Serial.print("Ton: ");
-  Serial.print(ontime);
-  Serial.println("us");
-
-  Serial.print("Toff: ");
-  Serial.print(offtime); 
-  Serial.println("us");   
-
-
-  Serial.print("Frequency: ");
-  Serial.print(freq);
-  Serial.println("Hz");
-
-  Serial.print("Duty: ");
-  Serial.print(duty); 
-  Serial.print('%'); 
-  delay(1000);
-  
-  for (int i = 0; i < 50; i++){
-    Serial.println();
-  }
-}
